Tests/LibWeb/Text/expected/vv-resize-on-pinch-to-zoom.txt

vv resize: width=533.3333333333334, height=400, scale=1.5, offsetLeft=33.328125, offsetTop=33.328125
